move Exception to common exceptions
